feat: Add diagnostic and verification scripts for runtime_core
- Introduced diagnostic.py for checking runtime_core installation, WASM build, and backend module imports.
- Added pyrightconfig.json for type checking configuration.
- Created test_e2e.py for end-to-end integration testing of FractalSync, covering imports, feature extraction, orbit synthesis, model initialization, and visual metrics.
- Implemented test_integration.py for comprehensive runtime-core integration tests, including WASM artifact verification.
- Developed verify_installation.py to confirm runtime_core installation and functionality.
- Added verify_runtime_core.py for quick checks on runtime_core and its components.
- Updated wasm-orbit Cargo.toml to include runtime_core as a dependency and added getrandom for JS compatibility.
- Enhanced wasm-orbit/src/lib.rs with new features for orbit state management and synthesis, including constants exposure to JavaScript.

+++ backend/src/visual_metrics.py
@@ -53,9 +53,9 @@ class VisualMetrics:
             )
 
         # Brightness distribution
-        metrics["brightness_mean"] = np.mean(gray)
-        metrics["brightness_std"] = np.std(gray)
-        metrics["brightness_range"] = np.max(gray) - np.min(gray)
+        metrics["brightness_mean"] = float(np.mean(gray))
+        metrics["brightness_std"] = float(np.std(gray))
+        metrics["brightness_range"] = float(np.max(gray) - np.min(gray))
 
         # Temporal change rate
         if prev_image is not None:
@@ -128,7 +128,7 @@ class VisualMetrics:
         local_var = cv2.filter2D((gray - local_mean) ** 2, -1, kernel)
 
         # Average variance (lower is more uniform)
-        avg_variance = np.mean(local_var)
+        avg_variance = float(np.mean(local_var))
 
         # Convert to uniformity score [0, 1] (inverse relationship)
         uniformity = 1.0 / (1.0 + avg_variance * 10.0)
